Process an entire batch at once - lets us call batch predict.  Still need to batch update as well
Process batch using single batches the same as the previous method

Uses the IncompleteParse namedtuple

Also, adds all the updates and runs them at once.  The numbers are apparently slightly different from separate updates in a batch because of 32 bit precision - separate updates gets 10520.843908309937 for the first batch vs 10520.84375 as a combined update.